+++ src/OpenTelemetry.Instrumentation.Http/CHANGELOG.md
@@ -11,6 +11,9 @@
   `HttpWebRequest` calls.
   ([#5000](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5000))
 
+* Updated OpenTelemetry core component version(s) to `1.18.0`.
+  ([#5022](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5022))
+
 ## 1.17.0
 
 Released 2026-Jul-17